+++ meo/src/main/java/com/mfano/meo/repositories/UserStatusRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.mfano.meo.models.UserStatus;
 
 @Repository
-public interface UserStatusRepository extends JpaRepository<UserStatus, Long>{
+public interface UserStatusRepository extends JpaRepository<UserStatus, Integer>{
 
 }
